+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 2.1.2 (Unreleased)
 ------------------
 
--   No changes yet.
+-   Do not make pe skymap public automatically before approved.
 
 2.1.1 "Lone Island Mountain Devil" (2023-05-29)
 -----------------------------------------------