+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 -   Do not make pe skymap public automatically before approved.
 
+-   Make image created when omegascan fails smaller.
+
 2.1.1 "Lone Island Mountain Devil" (2023-05-29)
 -----------------------------------------------