+++ CHANGES.rst
@@ -8,6 +8,9 @@ Changelog
 
 -   Make image created when omegascan fails smaller.
 
+-   Remove "overall state of detectors" line from detchar vector table,
+    since without Virgo being online it will always be "bad."
+
 2.1.1 "Lone Island Mountain Devil" (2023-05-29)
 -----------------------------------------------