+++ CHANGES.rst
@@ -13,6 +13,8 @@ Changelog
 
 -   Check the iDQ OK vector before checking the value of iDQ timeseries.
 
+-   Upgrade bilby_pipe to v1.1.1
+
 2.1.1 "Lone Island Mountain Devil" (2023-05-29)
 -----------------------------------------------